backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
celery[redis]
SQLAlchemy
alembic
pydantic
requests
boto3
